requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dateutil==2.8.2
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytesseract==0.3.13
opencv-python==4.10.0.84
numpy>=1.24.3,<2.0.0
langchain==0.1.0
langchain-openai==0.0.5
openai==1.12.0
python-dotenv==1.0.0
bleach==6.1.0